+++ test/test_database.rb
@@ -97,26 +97,6 @@ module SQLite3
       assert_equal [1], @db.get_first_row("SELECT 1")
     end
 
-    def test_get_first_row_with_type_translation_and_hash_results
-      @db.results_as_hash = true
-      capture_io do # hush translation deprecation warnings
-        @db.type_translation = true
-        assert_equal({"1" => 1}, @db.get_first_row("SELECT 1"))
-      end
-    end
-
-    def test_execute_with_type_translation_and_hash
-      rows = []
-      @db.results_as_hash = true
-
-      capture_io do # hush translation deprecation warnings
-        @db.type_translation = true
-        @db.execute("SELECT 1") { |row| rows << row }
-      end
-
-      assert_equal({"1" => 1}, rows.first)
-    end
-
     def test_encoding
       assert @db.encoding, "database has encoding"
     end