+++ test/test_integration_pending.rb
@@ -74,4 +74,40 @@ class IntegrationPendingTestCase < SQLite3::TestCase
 
     assert_operator time.real * 1000, :>=, 1000
   end
+
+  def test_busy_handler_timeout_releases_gvl
+    work = []
+
+    Thread.new do
+      loop do
+        sleep 0.1
+        work << "."
+      end
+    end
+    sleep 1
+
+    @db.busy_handler_timeout = 1000
+    busy = Mutex.new
+    busy.lock
+
+    t = Thread.new do
+      db2 = SQLite3::Database.open("test.db")
+      db2.transaction(:exclusive) do
+        busy.lock
+      end
+    ensure
+      db2&.close
+    end
+    sleep 1
+
+    work << "|"
+    assert_raises(SQLite3::BusyException) do
+      @db.execute "insert into foo (b) values ( 'from 2' )"
+    end
+
+    busy.unlock
+    t.join
+
+    assert_operator work.size - work.find_index("|"), :>, 3
+  end
 end